augur/frontend: fix JS imports for renamed files
augur-app.jsx → app.jsx and augur_client.js → client.js. The plan was
purely a rename, but the JS internal imports also needed updating.
