+++ augur/frontend/main.jsx
@@ -1,7 +1,7 @@
 import React from "react";
 import { createRoot } from "react-dom/client";
 
-import AugurApp from "./augur-app.jsx";
+import AugurApp from "./app.jsx";
 
 const root = createRoot(document.getElementById("root"));
 root.render(<AugurApp />);
